Add CRF support
This patch adds the support for Clock Reference Format (CRF) specified
in chapter 10 from the IEEE 1722-2016 spec.

CRF is based on AVTP CRF PDU so this patch introduces the 'struct
avtp_crf_pdu' which is defined in its own header, as it's specific to
this protocol.

The public header avtp_crf.h defines the CRF fields, their possible
values, and get/set APIs to manipulate these fields.

The actual timestamps will be part of the payload of this struct and
the best way to handle them is left for applications.

Signed-off-by: Rajvi Jingar <rajvi.jingar@intel.com>