examples: CVF Talker example
This patch adds the examples/cvf-talker.c file that provides an
example of how to use libavtp CVF APIs. The example implements a very
simple CVF talker, that reads H.264 byte-stream from standard input, and
creates and sends AVTP CVF packets via network.

The easiest way to use this example is to combine it with a GStreamer
pipeline that generates H.264 data and sends to stdout.

The cvf-talker.c file has an example of such pipeline.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>