examples: IEC 61883/IIDC MPEG-TS talker example
This patch adds the examples/ieciidc-talker.c file that provides an
example of how to use libavtp IEC 61883/IIDC APIs. The example implements
a very simple talker, that reads MPEG-TS from standard input and creates
and send AVTP IEC 61883/IIDC packets via network.

The easiest way to use this example is to combine it with a GStreamer
pipeline that generates MPEG-TS data and send to stdout.

The ieciidc-talker.c file has an example of such pipeline.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>