examples: IEC 61883/IIDC MPEG-TS listener example
This patch adds the examples/ieciidc-listener.c file that provides an
example of how to use libavtp IEC 61883/IIDC APIs. The example
implements a very simple listener, which reads AVTP IEC 61883/IIDC
packets from the network and sends an MPEG-TS to standard output.

The easiest way to use this example is to combine it with a GStreamer
pipeline that reads the MPEG-TS from standard input and shows it on an X
display.

The ieciidc-listener.c file has an example of such pipeline.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>