examples: CVF Listener example
This patch adds the examples/cvf-listener.c file that provides an
example of how to use libavtp CVF APIs. The example implements a very
simple CVF listener, which reads AVTP CVF packets from the network, and
sends an H.264 byte-stream to standard output.

The easiest way to use this example is to combine it with a GStreamer
pipeline that reads the H.264 byte-stream from standard input and shows
it on an X display.

The cvf-listener.c file has an example of such pipeline.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>